Remove machine controllers
