+++ pkg/controllers/machine/disruption/emptiness.go
@@ -29,10 +29,11 @@ import (
 	"github.com/prometheus/client_golang/prometheus"
 
 	"github.com/aws/karpenter-core/pkg/apis/v1alpha5"
+	"github.com/aws/karpenter-core/pkg/apis/v1beta1"
 	"github.com/aws/karpenter-core/pkg/controllers/state"
 	"github.com/aws/karpenter-core/pkg/metrics"
-	machineutil "github.com/aws/karpenter-core/pkg/utils/machine"
 	"github.com/aws/karpenter-core/pkg/utils/node"
+	nodeclaimutil "github.com/aws/karpenter-core/pkg/utils/nodeclaim"
 )
 
 // Emptiness is a machine sub-controller that adds or removes status conditions on empty machines based on TTLSecondsAfterEmpty
@@ -43,34 +44,34 @@ type Emptiness struct {
 }
 
 //nolint:gocyclo
-func (e *Emptiness) Reconcile(ctx context.Context, provisioner *v1alpha5.Provisioner, machine *v1alpha5.Machine) (reconcile.Result, error) {
-	hasEmptyCondition := machine.StatusConditions().GetCondition(v1alpha5.MachineEmpty) != nil
+func (e *Emptiness) Reconcile(ctx context.Context, nodePool *v1beta1.NodePool, nodeClaim *v1beta1.NodeClaim) (reconcile.Result, error) {
+	hasEmptyCondition := nodeClaim.StatusConditions().GetCondition(v1beta1.NodeEmpty) != nil
 
-	// From here there are three scenarios to handle:
+	// From here there are a few scenarios to handle:
 	// 1. If TTLSecondsAfterEmpty is not configured, remove the emptiness status condition
-	if provisioner.Spec.TTLSecondsAfterEmpty == nil {
+	if nodePool.Spec.Deprovisioning.EmptinessTTL == nil {
 		if hasEmptyCondition {
-			_ = machine.StatusConditions().ClearCondition(v1alpha5.MachineEmpty)
-			logging.FromContext(ctx).Debugf("removing emptiness status condition from machine since emptiness is disabled")
+			_ = nodeClaim.StatusConditions().ClearCondition(v1beta1.NodeEmpty)
+			logging.FromContext(ctx).Debugf("removing emptiness status condition, emptiness is disabled")
 		}
 		return reconcile.Result{}, nil
 	}
-	// 2. If Machine is not initialized, remove the emptiness status condition
-	if initCond := machine.StatusConditions().GetCondition(v1alpha5.MachineInitialized); initCond == nil || initCond.IsFalse() {
+	// 2. If NodeClaim is not initialized, remove the emptiness status condition
+	if initCond := nodeClaim.StatusConditions().GetCondition(v1beta1.NodeInitialized); initCond == nil || initCond.IsFalse() {
 		if hasEmptyCondition {
-			_ = machine.StatusConditions().ClearCondition(v1alpha5.MachineEmpty)
-			logging.FromContext(ctx).Debugf("removing emptiness status condition from machine since machine isn't initialized")
+			_ = nodeClaim.StatusConditions().ClearCondition(v1beta1.NodeEmpty)
+			logging.FromContext(ctx).Debugf("removing emptiness status condition, isn't initialized")
 		}
 		return reconcile.Result{}, nil
 	}
 	// Get the node to check for pods scheduled to it
-	n, err := machineutil.NodeForMachine(ctx, e.kubeClient, machine)
+	n, err := nodeclaimutil.NodeForNodeClaim(ctx, e.kubeClient, nodeClaim)
 	if err != nil {
-		// 3. If Machine -> Node mapping doesn't exist, remove the emptiness status condition
-		if machineutil.IsDuplicateNodeError(err) || machineutil.IsNodeNotFoundError(err) {
-			_ = machine.StatusConditions().ClearCondition(v1alpha5.MachineEmpty)
+		// 3. If Node mapping doesn't exist, remove the emptiness status condition
+		if nodeclaimutil.IsDuplicateNodeError(err) || nodeclaimutil.IsNodeNotFoundError(err) {
+			_ = nodeClaim.StatusConditions().ClearCondition(v1beta1.NodeEmpty)
 			if hasEmptyCondition {
-				logging.FromContext(ctx).Debugf("removing emptiness status condition from machine since machine doesn't have a single node mapping")
+				logging.FromContext(ctx).Debugf("removing emptiness status condition, doesn't have a single node mapping")
 			}
 			return reconcile.Result{}, nil
 		}
@@ -81,9 +82,9 @@ func (e *Emptiness) Reconcile(ctx context.Context, provisioner *v1alpha5.Provisi
 	// nomination time ends since we don't watch node nomination events
 	// 4. If the Node is nominated for pods to schedule to it, remove the emptiness status condition
 	if e.cluster.IsNodeNominated(n.Name) {
-		_ = machine.StatusConditions().ClearCondition(v1alpha5.MachineEmpty)
+		_ = nodeClaim.StatusConditions().ClearCondition(v1beta1.NodeEmpty)
 		if hasEmptyCondition {
-			logging.FromContext(ctx).Debugf("removing emptiness status condition from machine since machine is nominated for pods")
+			logging.FromContext(ctx).Debugf("removing emptiness status condition, is nominated for pods")
 		}
 		return reconcile.Result{RequeueAfter: time.Second * 30}, nil
 	}
@@ -93,20 +94,20 @@ func (e *Emptiness) Reconcile(ctx context.Context, provisioner *v1alpha5.Provisi
 	}
 	// 5. If there are pods that are actively scheduled to the Node, remove the emptiness status condition
 	if len(pods) > 0 {
-		_ = machine.StatusConditions().ClearCondition(v1alpha5.MachineEmpty)
+		_ = nodeClaim.StatusConditions().ClearCondition(v1beta1.NodeEmpty)
 		if hasEmptyCondition {
-			logging.FromContext(ctx).Debugf("removing emptiness status condition from node")
+			logging.FromContext(ctx).Debugf("removing emptiness status condition, not empty")
 		}
 		return reconcile.Result{}, nil
 	}
 	// 6. Otherwise, add the emptiness status condition
-	machine.StatusConditions().SetCondition(apis.Condition{
-		Type:     v1alpha5.MachineEmpty,
+	nodeClaim.StatusConditions().SetCondition(apis.Condition{
+		Type:     v1beta1.NodeEmpty,
 		Status:   v1.ConditionTrue,
 		Severity: apis.ConditionSeverityWarning,
 	})
 	if !hasEmptyCondition {
-		logging.FromContext(ctx).Debugf("marking machine as empty")
+		logging.FromContext(ctx).Debugf("marking empty")
 		metrics.MachinesDisruptedCounter.With(prometheus.Labels{
 			metrics.TypeLabel:        metrics.EmptinessReason,
 			metrics.ProvisionerLabel: machine.Labels[v1alpha5.ProvisionerNameLabelKey],
