--- pkg/controllers/nodeclaim/lifecycle/initialization.go
@@ -1,143 +1,0 @@
-/*
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package lifecycle
-
-import (
-	"context"
-	"fmt"
-
-	"github.com/samber/lo"
-	v1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/api/equality"
-	"knative.dev/pkg/logging"
-	"sigs.k8s.io/controller-runtime/pkg/client"
-	"sigs.k8s.io/controller-runtime/pkg/reconcile"
-
-	"github.com/aws/karpenter-core/pkg/apis/v1beta1"
-	"github.com/aws/karpenter-core/pkg/scheduling"
-	nodeutil "github.com/aws/karpenter-core/pkg/utils/node"
-	nodeclaimutil "github.com/aws/karpenter-core/pkg/utils/nodeclaim"
-	"github.com/aws/karpenter-core/pkg/utils/resources"
-)
-
-type Initialization struct {
-	kubeClient client.Client
-}
-
-// Reconcile checks for initialization based on if:
-// a) its current status is set to Ready
-// b) all the startup taints have been removed from the node
-// c) all extended resources have been registered
-// This method handles both nil provisioners and nodes without extended resources gracefully.
-func (i *Initialization) Reconcile(ctx context.Context, nodeClaim *v1beta1.NodeClaim) (reconcile.Result, error) {
-	if nodeClaim.StatusConditions().GetCondition(v1beta1.NodeInitialized).IsTrue() {
-		return reconcile.Result{}, nil
-	}
-	if !nodeClaim.StatusConditions().GetCondition(v1beta1.NodeLaunched).IsTrue() {
-		nodeClaim.StatusConditions().MarkFalse(v1beta1.NodeInitialized, "NotLaunched", "Node node launched")
-		return reconcile.Result{}, nil
-	}
-	ctx = logging.WithLogger(ctx, logging.FromContext(ctx).With("provider-id", nodeClaim.Status.ProviderID))
-	node, err := nodeclaimutil.NodeForNodeClaim(ctx, i.kubeClient, nodeClaim)
-	if err != nil {
-		nodeClaim.StatusConditions().MarkFalse(v1beta1.NodeInitialized, "NodeNotFound", "Node not registered with cluster")
-		return reconcile.Result{}, nil //nolint:nilerr
-	}
-	ctx = logging.WithLogger(ctx, logging.FromContext(ctx).With("node", node.Name))
-	if nodeutil.GetCondition(node, v1.NodeReady).Status != v1.ConditionTrue {
-		nodeClaim.StatusConditions().MarkFalse(v1beta1.NodeInitialized, "NodeNotReady", "Node status is NotReady")
-		return reconcile.Result{}, nil
-	}
-	if taint, ok := StartupTaintsRemoved(node, nodeClaim); !ok {
-		nodeClaim.StatusConditions().MarkFalse(v1beta1.NodeInitialized, "StartupTaintsExist", "StartupTaint %q still exists", formatTaint(taint))
-		return reconcile.Result{}, nil
-	}
-	if taint, ok := KnownEphemeralTaintsRemoved(node); !ok {
-		nodeClaim.StatusConditions().MarkFalse(v1beta1.NodeInitialized, "KnownEphemeralTaintsExist", "KnownEphemeralTaint %q still exists", formatTaint(taint))
-		return reconcile.Result{}, nil
-	}
-	if name, ok := RequestedResourcesRegistered(node, nodeClaim); !ok {
-		nodeClaim.StatusConditions().MarkFalse(v1beta1.NodeInitialized, "ResourceNotRegistered", "Resource %q was requested but not registered", name)
-		return reconcile.Result{}, nil
-	}
-	stored := node.DeepCopy()
-	node.Labels = lo.Assign(node.Labels, map[string]string{v1beta1.NodeInitializedLabelKey: "true"})
-	if !equality.Semantic.DeepEqual(stored, node) {
-		if err = i.kubeClient.Patch(ctx, node, client.MergeFrom(stored)); err != nil {
-			return reconcile.Result{}, err
-		}
-	}
-	logging.FromContext(ctx).Debugf("initialized %s", lo.Ternary(nodeClaim.IsMachine, "machine", "nodeclaim"))
-	nodeClaim.StatusConditions().MarkTrue(v1beta1.NodeInitialized)
-	nodeclaimutil.InitializedCounter(nodeClaim).Inc()
-	return reconcile.Result{}, nil
-}
-
-// KnownEphemeralTaintsRemoved validates whether all the ephemeral taints are removed
-func KnownEphemeralTaintsRemoved(node *v1.Node) (*v1.Taint, bool) {
-	for _, knownTaint := range scheduling.KnownEphemeralTaints {
-		// if the node still has a known ephemeral taint applied, it's not ready
-		for i := range node.Spec.Taints {
-			if knownTaint.MatchTaint(&node.Spec.Taints[i]) {
-				return &node.Spec.Taints[i], false
-			}
-		}
-	}
-	return nil, true
-}
-
-// StartupTaintsRemoved returns true if there are no startup taints registered for the provisioner, or if all startup
-// taints have been removed from the node
-func StartupTaintsRemoved(node *v1.Node, nodeClaim *v1beta1.NodeClaim) (*v1.Taint, bool) {
-	if nodeClaim != nil {
-		for _, startupTaint := range nodeClaim.Spec.StartupTaints {
-			for i := range node.Spec.Taints {
-				// if the node still has a startup taint applied, it's not ready
-				if startupTaint.MatchTaint(&node.Spec.Taints[i]) {
-					return &node.Spec.Taints[i], false
-				}
-			}
-		}
-	}
-	return nil, true
-}
-
-// RequestedResourcesRegistered returns true if there are no extended resources on the node, or they have all been
-// registered by device plugins
-func RequestedResourcesRegistered(node *v1.Node, nodeClaim *v1beta1.NodeClaim) (v1.ResourceName, bool) {
-	for resourceName, quantity := range nodeClaim.Spec.Resources.Requests {
-		if quantity.IsZero() {
-			continue
-		}
-		// kubelet will zero out both the capacity and allocatable for an extended resource on startup, so if our
-		// annotation says the resource should be there, but it's zero'd in both then the device plugin hasn't
-		// registered it yet.
-		// We wait on allocatable since this is the value that is used in scheduling
-		if resources.IsZero(node.Status.Allocatable[resourceName]) {
-			return resourceName, false
-		}
-	}
-	return "", true
-}
-
-func formatTaint(taint *v1.Taint) string {
-	if taint == nil {
-		return "<nil>"
-	}
-	if taint.Value == "" {
-		return fmt.Sprintf("%s:%s", taint.Key, taint.Effect)
-	}
-	return fmt.Sprintf("%s=%s:%s", taint.Key, taint.Value, taint.Effect)
-}
